ci: Add rustfmt
It will run rustfmt on the diffs and fail if it finds any
failures.

Note that this action does not work as intended with PRs
created with forked repositories (because of a github
limitation). But it will still fail and we can see the
failures in the logs instead of it being posted as a review
comment.

Sample failing action:

https://github.com/webmproject/CrabbyAvif/actions/runs/8053598990