ci: Ensure that avif.h is kept up to date
This file is used by C/C++ libraries that need to depend on
CrabbyAvif.

It is generated by using cbindgen and is checked into the
repository so that apps that don't use the cargo build system do
not have to build it separately.

This CI makes sure that the header file is in sync with any
changes to the rust code.