alpha: Remove incorrect alpha_premultiplied check
This check was added in PR #6. This function was sort of
mirroring the libavif C code and
`rgb::Image::alpha_premultiplied` is intended to be an input
parameter.

So returning `InvalidArgument` in `premultiply_alpha` when it
is set to `true` (potentially by the user) is not correct.

We can assume that this function will never be called twice
with the same frame (so that we will end up premultiplying
twice). If we really want to guard against this then i think
we should have a separate private variable keeping track of
whether alpha was already premultiplied as requested.

Ditto for unpremultiply_alpha

The newly added test fails without this PR.