crabbyavif: Add shorthand for as_ref and as_mut unwrapping
This is used frequently enough that calling as_ref().unwrap() and
as_mut().unwrap() can be replaced with a simpler trait function
unwrap_ref() and unwrap_mut() respectively.

I also considered naming the functions unref() and unmut() so that
they are shorter, but it felt like the term unref() could be
confused with unreference/dereference.

PiperOrigin-RevId: 615084910