crabbyavif, reformat: Fix half float conversion
In rust code, we cannot cast f32 to u32 using the `as` operator.

What we really want here is to transmute the underlying bits of
the f32 value and interpret that as u32 (equivalent to
reinterpret_cast in c++). The safe rust way to do it is to
convert it into a byte array and convert the byte array into u32.

An alternative is to call std::mem::transmute, but that is
unnecessary use of unsafe rust.

PiperOrigin-RevId: 615136885