decoder: Limit seq hdr search by first sample size
When trying to harvest the sequence header, limit the search to
the size of the first sample or 64 whichever is lower.

Since we create a slice unconditionally in partial_data, the
current code panics if the first sample's size is < 64. For extra
panic safety, also add a check before taking a slice in partial_data
so that we fail gracefully in cases like this.

PiperOrigin-RevId: 626532217