stream: Do not fail on invalid utf8 strings
When parsing the MP4 boxes, we encounter some utf8 strings. We don't
really have to verify if they are valid utf8 or not. We can continue
with the parsing as long as we find a null terminator.

For strings that we care about, like auxiliary URN, we know that it
will be valid utf8 anyways. So if we are unable to parsed the string
as utf8, simply assume it to be "" rather than failing.

PiperOrigin-RevId: 626535500