Fix size 0 handling
Introduce an enum to handle size 0. If the raw size occurring
in the file is 0, then the newly added enum field will be used
to mean that the box goes on until the end of the input stream.

Otherwise, fixed size is used.

Similar to https://github.com/AOMediaCodec/libavif/pull/2152

PiperOrigin-RevId: 631451055