reformat: Implement scale for non-pointer inputs
This case is not frequent but possible theoretically if an image
satisfies the following:
* Has limited range alpha.
* Needs scaling.
* Does not have grid.

In this case, we would convert the raw pointer into a managed
buffer when converting the alpha from limited to full range, and
then we would try to scale the managed buffer.

While this makes two copies of the buffer (one for alpha range
conversion and other for scaling), such images are not common
and performance is not a concern.

PiperOrigin-RevId: 632195724