pixels: Validate pointer and size in PointerSlice::create
slice::from_raw_parts requires the pointer to be aligned and
non-null, and the total size of the slice not to exceed
`isize::MAX`.

Perform those validations and change the entire call stack to
return an `AvifResult` instead. In the C API wrapper, we
simply convert the error into None and the underlying
functions will then fail gracefully.

PiperOrigin-RevId: 665098025