Use correct list of categories as requested
In the current code, only `decode_tiles` checks for the
ImageContentType value. Some of the other functions should also
check for it to reduce the work they are doing (prepare_samples
and create_codecs).

While create_codecs is cheap when using dav1d directly, it is
expensive in Android as creating a mediacodec is not a cheap
operation. So we should only create codecs which will be used
for decoding based on the ImageContentType.

PiperOrigin-RevId: 691386057