image: Use u32 math in copy_from_tile
The current code converts a lot of u32 variables into u64 and then
into usize. It is unnecessary as all of it will fit into u32
directly and we can go to usize from u32 directly. Remove all the
u64 conversions in copy_from_tile.

Also remove a stale TODO.

PiperOrigin-RevId: 693031602