mp4box: Support truncated ftyp box when peeking
In `peek_compatible_file_type`, try to do the best effort to
determine if a file is a valid image or not by accepting data
that does not contain the entire ftyp box. The new code reads
as many brands as possible and return true if any of the brands
matches the requirements.

Also update the test to test the new code path.

PiperOrigin-RevId: 693469534