capi: Fix alpha plane scaling
The scale function only works for one category at a time. When
it processes one category, it sets the width and height of the
image to that of of the scaled image. So if scale is called again
with the same image for the alpha category, it will be a no-op.

Restore the width and height before calling scale for the alpha
plane.

PiperOrigin-RevId: 697796689