capi: Do not copy entire plane in avifImageCopy
In the current implementation of avifImageCopy, we do a memcpy
of the entire plane (of size plane_height * row_bytes). This
is problematic in the following practical case:

1) Let's say we have |image1| of size 100x100 (with stride 104).
2) |image2| is a view of |image1| pixels but only for the last 20x100
   pixels.
3) Invoke copy on |image2|.

When trying to copy the last row of |image2| we will end up accessing
invalid pixels if we copy the whole plane (since we will go past
the actual stride).

Instead, copy the pixels row by row and only copy |width| pixels
in each row.

Newly added test will fail under ASAN without the fix.

PiperOrigin-RevId: 719447945