decoder: Add a function to decode grid images
Add a function to the Decoder trait that will decode grid images
in one shot. This is helpful for async decoders which can decode
multiple frames at the same time.

This CL only adds the trait interface without any actual
implementations. Android mediacodec implementation of this trait
will come in a follow-up.

PiperOrigin-RevId: 737657674