example: Add encoder utility
The crabbyavif utility will now support AVIF encoding as well.
It decides the operating mode (encode or decode) based on the
file extensions.

The command line args are not well defined yet. So this can be
considered as a work in progress that will be improved and
documented over time once all features are added.

PiperOrigin-RevId: 739022046