scale: Support scaling P010 images
P010 images cannot be scaled with ScalePlane_12 since the U and V
planes are interleaved. libyuv doesn't have a function to do this
scaling directly. So convert the image into I010 (which has separate
U and V planes) and then scale each plane by calling ScalePlane_12.

There is a small nuance in the C-API code where we will have to
explicitly allocate the V plane's memory buffer when going through
this case.

PiperOrigin-RevId: 743809051