decoder: Support progressive decoding with partial data
Allow the decoder to return a progressive layer as long as it has
enough data to decoder that layer. The existing tests, unfortunately,
do not cover this scenario (as they always test the progressive state
when the whole data is available).

Add tests to mimic the partial data behavior (similar to how we
test incremental decoding).

PiperOrigin-RevId: 758734942