mp4box: Do not enforce version number in iinf
This check seems to come from libavif but i could not trace the
source of the libavif check either since its been there since the
first versrion [1].

The spec, however, does not say anywhere that version number cannot
be > 1. It has the following if condition:
```
if version == 0 {
  ...
} else {
  ...
}
```

So we should also use the same condition.

[1] https://github.com/AOMediaCodec/libavif/commit/444f051999b4178797f1724d157872f8805fe0b1

PiperOrigin-RevId: 758774944