capi: Support setting codec specific options
The format of the C API is similar to that of libavif:
c:key or color:key for color only options.
a:key or alpha:key for alpha only options.
no prefix for options that apply for all planes.

In addition we also support,
g:key or gainmap:key for gainmap only options.
PiperOrigin-RevId: 770364032