capi: Handle codec specific options before init
In the C API, the underlying rust encoder is initialized only on
the first call to avifEncoderAddImage*() functions. But
avifEncoderSetCodecSpecificOption can be called before or after
that. So handle both cases by storing them in a temporary set
if it is called before initialization and copy them over during
initialization.

PiperOrigin-RevId: 772092324