Revert "capi: Add a libavif compatible header file"
Expose another C/C++ header file which will provide the symbols
the same way it is exposed in libavif. This is intended to be a
drop-in replacement for libavif.

It has the following differences from avif.h:
* No namespace.
* Functions and constants can be used without the crabby_/CRABBY_
prefix.

If an app is only using crabbyavif, it can use avif_compat.h to
not re-write any of the...

PiperOrigin-RevId: 776316310