encoder: Add setting to allow byte exact container
Animated AVIF images have a few time fields (creationTime,
modificationTime) which are written to `now()` by default. This
can be problematic for apps that expect a byte exact output for
the same set of inputs.

Add an encoder setting to override these fields. A value of `0`
means `now()` will be used, otherwise the provided value will be
used.

PiperOrigin-RevId: 814354264