mediacodec: Set EOS for single images
This allows certain type of HEIC images to be decoded by
MediaCodec.

MediaCodec delays frame output for certain type of HEIC images. To
force it to return a frame, EOS has to be set (if the codec is used
exactly for one image).

We already set this flag for grid images in get_next_image_grid_impl.

Fixes github issue #692

PiperOrigin-RevId: 827718355